requirements.txt
firecrawl-py
pydantic
pandas
python-dotenv
